third_party/WebKit/LayoutTests/http/tests/websocket/handshake-error-expected.txt

CONSOLE ERROR: line 20: WebSocket connection to 'ws://127.0.0.1:8880/handshake-error' failed: Error during WebSocket handshake: net::ERR_INVALID_HTTP_RESPONSE
Handshake error test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
Closed.
PASS ws.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
